updated file structure and applied review suggestions
