+++ apps/examples/package.json
@@ -13,7 +13,7 @@
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@mui/icons-material": "^6.4.0",
-    "@mui/joy": "^5.0.0-beta.51",
+    "@mui/joy": "5.0.0-beta.52",
     "@mui/material": "^6.3.1",
     "formik": "^2.4.6",
     "formik-form-builder": "workspace:*",
